+++ src/Compute/Compute.ts
@@ -4,7 +4,7 @@ import { clearWatcher } from '../helpers/clearWatcher'
 import { destroyWatchers } from '../helpers/destroyWatchers'
 import { watchWithScope } from '../helpers/watchWithScope'
 import { Observable } from '../Observable'
-import type { Destructor, Observer, Selector, Watcher } from '../types'
+import type { Destructor, Observer, Reaction, Watcher } from '../types'
 import { shiftSet } from '../utils/shiftSet'
 
 /* queue */
@@ -146,10 +146,16 @@ export class Compute<V = unknown> extends Observable<V> implements Observer {
     return this.childrenObservers
   }
 
-  constructor (watcher: Selector<V>, freeParent?: boolean, fireImmediately?: boolean)
-  /** @deprecated `update` argument is deprecated, use `Selector` */
-  constructor (watcher: Watcher<V>, freeParent?: boolean, fireImmediately?: boolean)
-  constructor (readonly watcher: Watcher<V> | Selector<V>, freeParent?: boolean, fireImmediately?: boolean) {
+  // TODO: remove in major release
+  /** @deprecated Use `reaction` */
+  get watcher () {
+    return this.reaction
+  }
+
+  constructor (reaction: Reaction<V>, freeParent?: boolean, fireImmediately?: boolean)
+  /** @deprecated `update` argument is deprecated, use `Reaction` */
+  constructor (reaction: Watcher<V>, freeParent?: boolean, fireImmediately?: boolean)
+  constructor (readonly reaction: Watcher<V> | Reaction<V>, freeParent?: boolean, fireImmediately?: boolean) {
     super()
 
     if (!freeParent) {
@@ -182,7 +188,7 @@ export class Compute<V = unknown> extends Observable<V> implements Observer {
       this.invalid = false
 
       watchWithScope(this, () => {
-        const newValue = this.watcher(this.updated) // TODO: remove `this.updated` in major release
+        const newValue = this.reaction(this.updated) // TODO: remove `this.updated` in major release
         this.updated = true
 
         if (newValue !== this.raw) {